requirements.txt
transformers>=4.30.0
peft>=0.4.0
datasets>=2.12.0
accelerate>=0.20.0
wandb>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
timm>=0.9.0
omegaconf>=2.3.0
